webkit/data/layout_tests/platform/chromium-win/LayoutTests/media/video-append-source-expected.txt

EXPECTED (video.currentSrc == '') OK
EXPECTED (video.currentSrc == '') OK
EVENT(load)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.ogv') OK
END OF TEST

